% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanse_data.R
\name{cleanse_data}
\alias{cleanse_data}
\title{Replace outliers and missing values in tidy time series data.}
\usage{
cleanse_data(.data, variable, replace.missing = TRUE)
}
\arguments{
\item{.data}{A tsibble returned by \code{detect_outliers}}

\item{variable}{A column to find outliers}

\item{replace.missing}{If TRUE, it not only replaces outliers,
but also interpolates missing values. Defualt is set to TRUE.}
}
\description{
Replace outliers and missing values in
the output returned by \code{detect_outliers}
}
\examples{
library(fabletools)
data <- USAccDeaths
# Convert 20th data point as an outlier
data[20] <- data[20]*5
p <- data \%>\%
  tsibble::as_tsibble() \%>\%
  detect_outliers(variable = "value",
  cmbn_model = c("arima", "theta", "ave"),
  p_rate = 0.001)
altered_data <- p \%>\% cleanse_data(variable = "value")
altered_data \%>\% fabletools::autoplot(.altered)

}
