% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstable-meta.R
\name{is_outstable}
\alias{is_outstable}
\title{Test if the object is an outstable}
\usage{
is_outstable(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object inherits from the `outstable` class.
}
\description{
This function returns TRUE for
 outstable or subclasses thereof, and
 FALSE for all other objects.
 Outstable can be generated by \code{detect_outliers}
}
\examples{
library(fabletools)
data <- USAccDeaths
# Convert 20th data point as an outlier
data[20] <- data[20]*5
p <- data \%>\%
  tsibble::as_tsibble() \%>\%
  detect_outliers(variable = "value",
  cmbn_model = c("arima", "theta", "ave"),
  p_rate = 0.001)
is_outstable(p)

}
