% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstable-package.R
\docType{package}
\name{outstable-package}
\alias{outstable-package}
\title{outstable: Tidy Time Series Anomaly Detection}
\description{
The **outstable** package provides a framework for detecting
anomalies in tidy time series data. An anomaly is defined as an
observation that is predicted as very unlikely given the robust time
series forecast models. The algorithm works with tidy temporal data
provided by the tsibble package and produces an outstable, a tsibble
with flagged anomalies and their degree of outlierness. An approach
based on extreme value theory is applied to residual series in order to
calculate a data-driven anomalous threshold. The proposed framework
can also provide a cleansed tsibble that closely integrates with the
tidy forecasting workflow used in the fable package. A number of
different approaches are available for the data cleansing process.
}
\keyword{internal}
