% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cutoff.R
\name{set_cutoff}
\alias{set_cutoff}
\title{Set a threshold for outlier detection}
\usage{
set_cutoff(x, p_rate = 0.01, trials = 500)
}
\arguments{
\item{x}{A numerical vector}

\item{p_rate}{False positive rate. Default value is set to 0.001}

\item{trials}{Number of trials to generate the extreme value distirbution. Default value is set to 500.}
}
\value{
Returns a list of flagged anomalies and their degree of
 outlierness
}
\description{
This function forecasts a boundary
for the typical behaviour using a representative
sample of the typical behaviour derived from a
given vector, An approach based on extreme value
theory is used for this boundary prediction process.
}
\examples{
p <- c(rnorm(100),1000)
out <- outstable::set_cutoff(p, p_rate = 0.01)
print(out)

}
\references{
Clifton, D. A., Hugueny, S., & Tarassenko, L. (2011). Novelty detection with multivariate extreme value statistics.
Journal of signal processing systems, 65 (3),371-389.
}
