% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Flagged anomalies and their degree of outlierness}
\usage{
detect_outliers(
  .data,
  variable,
  cmbn_model = c("ets", "arima", "snaive", "naive", "croston", "ave", "neural", "lm",
    "theta", "fasster"),
  ...
)
}
\arguments{
\item{.data}{A tsibble.}

\item{variable}{A column to find outliers}

\item{cmbn_model}{A vector of models to compute combination forcast}

\item{...}{Other arguments passed to [fabletools::model()].}
}
\description{
Flagged anomalies and their degree of
 outlierness
}
\examples{
library(fabletools)
data <- USAccDeaths
# Convert 10th data point as an outlier
data[10] <- 700000
p <- data \%>\%
  tsibble::as_tsibble() \%>\%
  detect_outliers(variable = "value",
  cmbn_model = c("naive","snaive", "ave"))
p
}
